Add estimated ROI and cost to categorical market buy
