+++ app/src/Sidebar/components/CategoricalBuy/buy.scss
@@ -29,64 +29,8 @@
   }
 
   &-summary {
-    &-heading {
-      @include sidebar-entry;
-      padding: 1rem 1.2rem;
-      font-size: 1.3rem;
-      color: $shadeBlue;
-      border-bottom: none;
-    }
-
-    &-category {
-      @include sidebar-entry;
-      font-size: 1.5rem;
-      line-height: 1;
-      background-color: $grey;
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-      min-height: 4.6rem;
-      padding: 1rem 1.6rem 0;
-      flex-flow: row wrap;
-    }
-
-    .category-values {
-      margin: -1rem 0 0 auto;
-      flex: 0 1 auto;
-      font-size: 1.3rem;
-      letter-spacing: 0;
-      text-align: right;
-      line-height: 1;
-
-      &-positive {
-        color: $green;
-      }
-
-      &-negative {
-        color: $red;
-      }
-
-      .category-value {
-        text-align: right;
-        white-space: nowrap;
-        margin: 0;
-
-        &.value {
-          font-weight: $bold;
-        }
-
-        &.margin {
-          font-size: 80%;
-        }
-      }
-    }
-
-    .category-entries {
-      display: flex;
-      align-self: center;
-      flex-flow: row wrap;
-      flex: 1;
-    }
+    @include sidebar-entry;
+    display: block;
   }
 
   &-investment {
