+++ app/src/Sidebar/components/CategoricalBuy/index.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import PropTypes from "prop-types";
 import Decimal from "decimal.js-light";
 
@@ -10,14 +10,10 @@ import style from "./buy.scss";
 import useGlobalState from "hooks/useGlobalState";
 
 import OutcomeSelection from "./OutcomeSelection";
+import BuySummary from "./BuySummary";
 import AmountInput from "../Buy/AmountInput";
-import OutcomeCard from "components/OutcomeCard";
 import { zeroDecimal } from "utils/constants";
-import { formatCollateral } from "utils/formatting";
-import {
-  calcPositionGroups,
-  calcOutcomeTokenCounts
-} from "utils/position-groups";
+import { calcOutcomeTokenCounts } from "utils/position-groups";
 
 const { BN } = Web3.utils;
 
@@ -63,7 +59,6 @@ const Buy = ({
   const { marketProbabilities } = useGlobalState();
 
   const [investmentAmount, setInvestmentAmount] = useState("");
-  const [humanReadablePositions, setHumanReadablePositions] = useState(null);
   const [error, setError] = useState(null);
   useEffect(() => {
     if (stagedTransactionType !== "buy outcome tokens") return;
@@ -171,16 +166,6 @@ const Buy = ({
     account
   ]);
 
-  const [stagedTradePositionGroups, setStagedTradePositionGroups] = useState(
-    []
-  );
-  useEffect(() => {
-    setStagedTradePositionGroups(
-      stagedTradeAmounts &&
-        calcPositionGroups(markets, positions, stagedTradeAmounts)
-    );
-  }, [markets, positions, stagedTradeAmounts]);
-
   let problemText;
 
   if (!marketStage === "Closed") {
@@ -199,110 +184,6 @@ const Buy = ({
       );
     }
   }, [collateralBalance, collateral]);
-
-  useEffect(() => {
-    const hasConditional = (marketSelections || []).some(
-      ({ isAssumed }) => isAssumed
-    );
-    let humanReadablePositions = {
-      payOutWhen: {
-        title: "Pay out when:",
-        getGlue: () => "and",
-        getPrefix: () => "IF",
-        positions: []
-      },
-      loseInvestmentWhen: {
-        title: "Lose investment when:",
-        positions: [],
-        getGlue: () => (hasConditional ? "and" : "or"),
-        getPrefix: () => "IF"
-      },
-      refundWhen: {
-        title: "Refund when:",
-        positions: [],
-        getGlue: () => "or",
-        getPrefix: () => "IF"
-      }
-    };
-
-    (stagedTradePositionGroups || []).forEach(
-      ({ outcomeSet, runningAmount }) => {
-        let hasEnteredInvestment;
-
-        try {
-          const decimalInvest = Decimal(investmentAmount);
-          hasEnteredInvestment = decimalInvest.gt(0);
-        } catch (err) {
-          //
-        }
-
-        // all payouts
-        humanReadablePositions.payOutWhen.positions = outcomeSet;
-        humanReadablePositions.payOutWhen.runningAmount = hasEnteredInvestment
-          ? runningAmount
-          : zeroDecimal;
-
-        // all lose invests
-
-        // invert outcome sets
-        humanReadablePositions.loseInvestmentWhen.positions = outcomeSet.map(
-          selectedOutcome => {
-            if (selectedOutcome.outcomeIndex == -1) {
-              return selectedOutcome;
-            }
-
-            if (marketSelections[selectedOutcome.marketIndex].isAssumed) {
-              return {
-                ...selectedOutcome,
-                ...markets[selectedOutcome.marketIndex].outcomes[
-                  selectedOutcome.outcomeIndex
-                ]
-              };
-            }
-
-            return {
-              ...selectedOutcome,
-              ...markets[selectedOutcome.marketIndex].outcomes[
-                selectedOutcome.outcomeIndex == 0 ? 1 : 0
-              ],
-              outcomeIndex: selectedOutcome.outcomeIndex == 0 ? 1 : 0
-            };
-          }
-        );
-        humanReadablePositions.loseInvestmentWhen.runningAmount = Decimal(
-          hasEnteredInvestment ? investmentAmount : zeroDecimal
-        )
-          .neg()
-          .mul(Math.pow(10, collateral.decimals));
-        humanReadablePositions.loseInvestmentWhen.margin = Decimal(-1.0);
-
-        // refund when
-
-        // invert outcome sets
-        humanReadablePositions.refundWhen.positions = outcomeSet
-          .filter(outcome => marketSelections[outcome.marketIndex].isAssumed)
-          .map(outcome => {
-            if (outcome.outcomeIndex == -1) {
-              return outcome;
-            }
-
-            return {
-              ...outcome,
-              ...markets[outcome.marketIndex].outcomes[
-                outcome.outcomeIndex == 0 ? 1 : 0
-              ],
-              outcomeIndex: outcome.outcomeIndex == 0 ? 1 : 0
-            };
-          });
-        humanReadablePositions.refundWhen.runningAmount = Decimal(
-          hasEnteredInvestment ? investmentAmount : zeroDecimal
-        ).mul(Math.pow(10, collateral.decimals));
-        humanReadablePositions.refundWhen.margin = Decimal(1.0);
-      }
-    );
-
-    setHumanReadablePositions(humanReadablePositions);
-  }, [stagedTradePositionGroups]);
 
   return (
     <>
@@ -348,43 +229,18 @@ const Buy = ({
           }}
         />
       </div>
-      {/* <div className={cx("buy-summary")}>
-        {humanReadablePositions &&
-          [
-            humanReadablePositions.payOutWhen,
-            humanReadablePositions.refundWhen,
-            humanReadablePositions.loseInvestmentWhen
-          ]
-            .filter(category => category && category.positions.length)
-            .map(category => (
-              <Fragment key={category.title}>
-                <div className={cx("buy-summary-heading")}>
-                  {category.title}
-                </div>
-                <div className={cx("buy-summary-category")}>
-                  <div className={cx("category-entries")}>
-                    {category.positions.map(outcome => (
-                      <OutcomeCard
-                        key={`${outcome.marketIndex}-${outcome.outcomeIndex}`}
-                        glueType={category.getGlue()}
-                        // prefixType={category.getPrefix()}
-                        {...outcome}
-                      />
-                    ))}
-                  </div>
-                  <div className={cx("category-values")}>
-                    <p className={cx("category-value", "value")}>
-                      {formatCollateral(category.runningAmount, collateral)}
-                    </p>
-                    {/*<p className={cx("category-value", "margin")}>
-                      ({category.margin > 0 && "+"}
-                      {category.margin * 100}%)
-                      </p>*/}
-      {/*</div>
-                </div>
-              </Fragment>
-            ))}
-      </div> */}
+      <div className={cx("buy-summary")}>
+        <BuySummary
+          {...{
+            markets,
+            positions,
+            collateral,
+            stagedTradeAmounts,
+            marketSelections,
+            investmentAmount
+          }}
+        />
+      </div>
       <div className={cx("buy-confirm")}>
         <button
           className={cx("button")}
@@ -468,6 +324,7 @@ Buy.propTypes = {
   setStagedTransactionType: PropTypes.func.isRequired,
   ongoingTransactionType: PropTypes.string,
   asWrappedTransaction: PropTypes.func.isRequired,
+  setMarketSelections: PropTypes.func.isRequired,
   resetMarketSelections: PropTypes.func.isRequired,
   makeButtonSelectCallback: PropTypes.func.isRequired
 };
